+++ packages/orq-rc/src/orq_ai_sdk/models/filegetop.py
@@ -59,5 +59,5 @@ class FileGetResponseBody(BaseModel):
     workspace_id: str
     r"""The id of the resource"""
 
-    created: Optional[datetime] = parse_datetime("2025-07-11T12:52:35.845Z")
+    created: Optional[datetime] = parse_datetime("2025-07-14T14:55:51.541Z")
     r"""The date and time the resource was created"""