+++ packages/orq-rc/RELEASES.md
@@ -1979,3 +1979,13 @@ Based on:
 - [python v3.10.0-rc.6] packages/orq-rc
 ### Releases
 - [PyPI v3.10.0-rc.6] https://pypi.org/project/orq-ai-sdk/3.10.0-rc.6 - packages/orq-rc
+
+## 2025-07-14 14:57:21
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.581.1 (2.656.9) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v3.10.0-rc.7] packages/orq-rc
+### Releases
+- [PyPI v3.10.0-rc.7] https://pypi.org/project/orq-ai-sdk/3.10.0-rc.7 - packages/orq-rc